+++ deepinv/loss/sup.py
@@ -1,8 +1,10 @@
+from __future__ import annotations
 from typing import Union
 
 import torch
 from deepinv.loss.loss import Loss
 from deepinv.loss.metric.metric import Metric
+from deepinv.models.base import Reconstructor
 
 
 class SupLoss(Loss):
@@ -41,3 +43,80 @@ class SupLoss(Loss):
         :return: (:class:`torch.Tensor`) loss.
         """
         return self.metric(x_net, x)
+
+
+class ReducedResolutionLoss(SupLoss):
+    r"""
+    Reduced resolution loss for blur and downsampling problems.
+
+    The reduced resolution loss is defined as
+
+    .. math::
+
+        \frac{1}{n}\|y-\inverse{\forw{y}}\|^2
+
+    where :math:`\forw{y}` is the further-degraded measurement, and the measurement :math:`y` is used a supervisory signal.
+
+    .. warning::
+
+        This loss can only be used with physics that can be used to meaningfully further degrade the measurements
+        :math:`y`, such as blur or downsampling. The physics must be defined without an `img_size` so it can be applied
+        to the measurements :math:`y`.
+
+    .. info::
+
+        During training, consider using the `disable_train_metrics` option in :class:`deepinv.Trainer` to prevent a shape
+        mismatch during metric computation since the reduced resolution output will smaller than ground truth.
+
+    :param Metric, torch.nn.Module metric: metric used for computing data consistency,
+        which is set as the mean squared error by default.
+    """
+
+    def forward(self, x_net, y, *args, **kwargs):
+        r"""
+        Computes the reduced resolution loss.
+
+        :param torch.Tensor x_net: reconstructions.
+        :param torch.Tensor y: Measurements.
+        :param deepinv.physics.Physics physics: Forward operator associated with the measurements.
+        :param torch.nn.Module model: Reconstruction function.
+        :return: (:class:`torch.Tensor`) loss.
+        """
+        try:
+            return self.metric(x_net, y)
+        except BaseException as e:
+            raise RuntimeError(
+                f"Metric error. Check that the reconstruction (of shape {x_net.shape}) and y (of shape {y.shape}) can be used to calculate the metric. Full error:",
+                str(e),
+            )
+
+    def adapt_model(self, model: torch.nn.Module) -> ReducedResolutionModel:
+        if isinstance(model, self.ReducedResolutionModel):
+            return model
+        else:
+            return self.ReducedResolutionModel(model)
+
+    class ReducedResolutionModel(Reconstructor):
+        def __init__(self, model):
+            super().__init__()
+            self.model = model
+
+        def forward(self, y, physics, **kwargs):
+            with torch.set_grad_enabled(self.training):
+                if self.training:
+                    try:
+                        y_rr = physics.A(y)
+                    except BaseException as e:
+                        raise RuntimeError(
+                            "Physics error. Check that the used physics can be applied to y to generate a further degraded y. Full error:",
+                            str(e),
+                        )
+                    try:
+                        return self.model(y_rr, physics)
+                    except BaseException as e:
+                        raise RuntimeError(
+                            "Model error. Check that the model can be used with a reduced-resolution input physics.A(y). Full error:",
+                            str(e),
+                        )
+                else:
+                    return self.model(y, physics)