+++ deepinv/tests/test_loss.py
@@ -24,6 +24,7 @@ LOSSES = [
     "vortex",
     "ensure",
     "ensure_mri",
+    "reducedresolution",
 ]
 
 LIST_SURE = [
@@ -143,6 +144,8 @@ def choose_loss(loss_name, rng=None, imsize=None, device="cpu"):
                 T_i=dinv.transform.RandomNoise(), T_e=dinv.transform.Shift()
             )
         )
+    elif loss_name == "reducedresolution":
+        loss.append(dinv.loss.ReducedResolutionLoss())
     else:
         raise Exception("The loss doesnt exist")
 
@@ -319,15 +322,22 @@ def test_losses(
         physics = dinv.physics.MRI(**gen.step(), device=device)
         loss = dinv.loss.mri.ENSURELoss(0.01, gen, rng=rng)
         dataset = _dataset(physics, tmp_path, imsize, device)
+    elif loss_name == "reducedresolution":
+        imsize = (1, 16, 16)
+        physics = dinv.physics.Downsampling(filter=None, factor=2, device=device)
+        dataset = _dataset(physics, tmp_path, imsize, device)
 
     save_dir = tmp_path / "dataset"
     # choose backbone denoiser
     backbone = dinv.models.AutoEncoder(
         dim_input=imsize[0] * imsize[1] * imsize[2], dim_mid=64, dim_hid=16
-    ).to(device)
+    )
+
+    if loss_name == "reducedresolution":
+        backbone = dinv.models.UNet(1, 1, scales=2)
 
     # choose a reconstruction architecture
-    model = dinv.models.ArtifactRemoval(backbone, device=device)
+    model = dinv.models.ArtifactRemoval(backbone.to(device), device=device)
 
     # choose optimizer and scheduler
     epochs = 10
@@ -352,6 +362,7 @@ def test_losses(
         plot_images=(loss_name == LOSSES[0]),  # save time
         verbose=False,
         log_train_batch=(loss_name == "sup_log_train_batch"),
+        disable_train_metrics=(loss_name == "reducedresolution"),
     )
 
     # test the untrained model
@@ -666,3 +677,20 @@ def test_stacked_loss(device, imsize):
     loss_value = loss(x=x, y=y, x_net=x_net, physics=physics, model=f)
 
     assert loss_value > 0
+
+
+def test_reducedresolution_shapes(device):
+    loss = dinv.loss.ReducedResolutionLoss()
+    model = dinv.models.ArtifactRemoval(DummyModel(), device=device)
+    model = loss.adapt_model(model)
+    x = torch.rand(1, 1, 8, 8, device=device)
+    physics = dinv.physics.Downsampling(filter=None, factor=2, device=device)
+    y = physics(x)
+
+    model.eval()
+    x_hat_eval = model(y, physics)
+    assert x_hat_eval.shape == x.shape
+
+    model.train()
+    x_hat_train = model(y, physics)
+    assert x_hat_train.shape == y.shape